Add Cart Screen

+++ data/src/main/java/com/abhinash/data/model/PriceDto.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 internal class PriceDto(
     @SerializedName("currency") val currency: String,
-    @SerializedName("amount") val amount: Int
+    @SerializedName("amount") val amount: Double
 ) {
-    fun toDomain(): Price = Price(currency, (amount/100).toDouble()) // Convert to cents
+    fun toDomain(): Price = Price(currency, amount/100) // Convert to cents
 }
